refactor: project 컴포넌트 구조 개선
- 페이지와 탭 컴포넌트를 별도 폴더로 분리
  - pages/: 페이지 레벨 컴포넌트
  - tabs/: 프로젝트 상세 페이지의 탭 컴포넌트

- 파일 이동:
  - project-page-content.tsx → pages/
  - project-detail-content.tsx → pages/
  - project-interviews.tsx → tabs/
  - project-insights.tsx → tabs/
  - project-settings.tsx → tabs/
  - interview-detail.tsx → tabs/

- index 파일 추가로 export 정리
- 모든 import 경로 업데이트
