+++ app/projects/[id]/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { use } from 'react'
-import { ProjectDetailContent } from '@/components/project/project-detail-content'
+import { ProjectDetailContent } from '@/components/project/pages/project-detail-content'
 
 interface ProjectPageProps {
   params: Promise<{ id: string }>
