+++ components/project/pages/index.ts
@@ -1,0 +1,2 @@
+export { ProjectPageContent } from './project-page-content'
+export { ProjectDetailContent } from './project-detail-content'
