+++ components/project/tabs/index.ts
@@ -1,0 +1,4 @@
+export { default as ProjectInterviews } from './project-interviews'
+export { default as ProjectInsights } from './project-insights'
+export { default as ProjectSettings } from './project-settings'
+export { default as InterviewDetail } from './interview-detail'
